requirements.txt
distro==1.8.0
PySide6==6.5.0
python-rtmidi==1.4.9
tqdm==4.65.0
